+++ packages/bundle/src/createCognitiveServicesSpeechServicesPonyfillFactory.js
@@ -1,8 +1,8 @@
 import { AudioConfig } from 'microsoft-cognitiveservices-speech-sdk/distrib/lib/src/sdk/Audio/AudioConfig';
 import createPonyfill from 'web-speech-cognitive-services/lib/SpeechServices';
 
-async function resolveFunction(fnOrValue) {
-  return await (typeof fnOrValue === 'function' ? fnOrValue() : fnOrValue);
+function resolveFunction(fnOrValue) {
+  return typeof fnOrValue === 'function' ? fnOrValue() : fnOrValue;
 }
 
 export default function createCognitiveServicesSpeechServicesPonyfillFactory({
@@ -25,15 +25,15 @@ export default function createCognitiveServicesSpeechServicesPonyfillFactory({
     credentials = async () => {
       if (authorizationToken) {
         return {
-          authorizationToken: resolveFunction(authorizationToken),
+          authorizationToken: await resolveFunction(authorizationToken),
           region
         };
-      } else {
-        return {
-          region,
-          subscriptionKey: resolveFunction(subscriptionKey)
-        };
       }
+
+      return {
+        region,
+        subscriptionKey: await resolveFunction(subscriptionKey)
+      };
     };
   }
 
@@ -57,12 +57,12 @@ export default function createCognitiveServicesSpeechServicesPonyfillFactory({
     };
   }
 
-  return ({ referenceGrammarID }) => {
+  return ({ referenceGrammarID } = {}) => {
     const ponyfill = createPonyfill({
       audioConfig,
       credentials,
       enableTelemetry,
-      referenceGrammars: [`luis/${referenceGrammarID}-PRODUCTION`],
+      referenceGrammars: referenceGrammarID ? [`luis/${referenceGrammarID}-PRODUCTION`] : [],
       speechRecognitionEndpointId,
       speechSynthesisDeploymentId,
       speechSynthesisOutputFormat,